+++ src/Lucene.Net.CodeAnalysis.Dev/LuceneDev1002_FloatingPointArithmeticCSCodeAnalyzer.cs
@@ -1,4 +1,4 @@
-﻿using Lucene.Net.CodeAnalysis.Dev.Helpers;
+﻿using Lucene.Net.CodeAnalysis.Dev.Utility;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
